RFPL-2261: Try fix autocomplete suggestions not displayed
There is problem with Autcomplete.typeKeys(...) routine: suggestions list doesn't display, and no characted in input appears even they was typed by selenium. It used to work with selenium1, but seems that it is working differently in Se2, so it no longer work as expected.
Have to find another way how to type keys incrementaly and fire event triggering suggestions display/update)